perf(pubmed): optimize date extraction and reduce class size
- Refactor `PubMedPublicationTransformer` to delegate date extraction to a new helper class `PubMedDateHelper` in `src/bioetl/application/pipelines/pubmed/date_helper.py`.
- Encapsulate date validation patterns and month mapping in `PubMedDateHelper` to avoid redundant object creation.
- Reduce `PubMedPublicationTransformer` class size from 731 to ~545 lines, resolving architecture complexity violation.
- Add comprehensive docstrings and type hints to the new helper class.

This change improves maintainability and performance by isolating date logic and reducing object churn.
